+++ praw/reddit.py
@@ -251,6 +251,7 @@ class Reddit(object):
                     self.config.kinds['subreddit']: models.Subreddit,
                     'LabeledMulti': models.Multireddit,
                     'Listing': models.Listing,
+                    'LiveUpdate': models.LiveUpdate,
                     'LiveUpdateEvent': models.LiveThread,
                     'UserList': models.RedditorList,
                     'modaction': models.ModAction,