+++ tests/integration/models/reddit/test_live.py
@@ -11,3 +11,11 @@ class TestLiveThread(IntegrationTest):
         thread = LiveThread(self.reddit, 'ukaeu1ik4sw5')
         with self.recorder.use_cassette('TestLiveThread_test_init'):
             assert thread.title == 'reddit updates'
+
+    @mock.patch('time.sleep', return_value=None)
+    def test_updates(self, _):
+        thread = LiveThread(self.reddit, 'ukaeu1ik4sw5')
+        with self.recorder.use_cassette('TestLiveThread_test_updates'):
+            for update in thread.updates(limit=None):
+                assert update.thread == thread
+        assert update.body.startswith('Small change:')